fastapi-lpocr-app/requirements.txt
fastapi[standard]==0.113.0
pydantic==2.8.0
SQLAlchemy==2.0.35
asyncpg==0.29.0
redis==5.1.1
SQLAlchemy[asyncio]==2.0.35
psycopg2-binary==2.9.9
SQLAlchemy-Utils==0.41.2
passlib[bcrypt]==1.7.4
azure-storage-blob==12.23.1
python-dotenv==1.0.1
fastapi-mail==1.4.1
PyJWT==2.9.0